[CAS-483] Move BaseDecorator to separate file
