+++ stream-chat-android-ui-components/src/main/kotlin/io/getstream/chat/android/ui/messages/adapter/viewholder/decorator/Decorator.kt
@@ -1,62 +1,8 @@
 package io.getstream.chat.android.ui.messages.adapter.viewholder.decorator
 
 import com.getstream.sdk.chat.adapter.MessageListItem
-import io.getstream.chat.android.core.internal.exhaustive
 import io.getstream.chat.android.ui.messages.adapter.BaseMessageItemViewHolder
-import io.getstream.chat.android.ui.messages.adapter.viewholder.MessageDeletedViewHolder
-import io.getstream.chat.android.ui.messages.adapter.viewholder.MessagePlainTextViewHolder
-import io.getstream.chat.android.ui.messages.adapter.viewholder.OnlyFileAttachmentsViewHolder
-import io.getstream.chat.android.ui.messages.adapter.viewholder.OnlyMediaAttachmentsViewHolder
-import io.getstream.chat.android.ui.messages.adapter.viewholder.PlainTextWithFileAttachmentsViewHolder
-import io.getstream.chat.android.ui.messages.adapter.viewholder.PlainTextWithMediaAttachmentsViewHolder
 
 internal interface Decorator {
     fun <T : MessageListItem> decorate(viewHolder: BaseMessageItemViewHolder<T>, data: T)
-}
-
-internal abstract class BaseDecorator : Decorator {
-    final override fun <T : MessageListItem> decorate(viewHolder: BaseMessageItemViewHolder<T>, data: T) {
-        if (data !is MessageListItem.MessageItem) {
-            return
-        }
-        when (viewHolder) {
-            is MessageDeletedViewHolder -> decorateDeletedMessage(viewHolder, data)
-            is MessagePlainTextViewHolder -> decoratePlainTextMessage(viewHolder, data)
-            is OnlyMediaAttachmentsViewHolder -> decorateOnlyMediaAttachmentsMessage(viewHolder, data)
-            is PlainTextWithMediaAttachmentsViewHolder -> decoratePlainTextWithMediaAttachmentsMessage(viewHolder, data)
-            is OnlyFileAttachmentsViewHolder -> decorateOnlyFileAttachmentsMessage(viewHolder, data)
-            is PlainTextWithFileAttachmentsViewHolder -> decoratePlainTextWithFileAttachmentsMessage(viewHolder, data)
-            else -> Unit
-        }.exhaustive
-    }
-
-    protected abstract fun decoratePlainTextWithFileAttachmentsMessage(
-        viewHolder: PlainTextWithFileAttachmentsViewHolder,
-        data: MessageListItem.MessageItem
-    )
-
-    protected abstract fun decorateOnlyFileAttachmentsMessage(
-        viewHolder: OnlyFileAttachmentsViewHolder,
-        data: MessageListItem.MessageItem
-    )
-
-    protected abstract fun decoratePlainTextWithMediaAttachmentsMessage(
-        viewHolder: PlainTextWithMediaAttachmentsViewHolder,
-        data: MessageListItem.MessageItem
-    )
-
-    protected abstract fun decorateOnlyMediaAttachmentsMessage(
-        viewHolder: OnlyMediaAttachmentsViewHolder,
-        data: MessageListItem.MessageItem
-    )
-
-    protected abstract fun decoratePlainTextMessage(
-        viewHolder: MessagePlainTextViewHolder,
-        data: MessageListItem.MessageItem
-    )
-
-    protected abstract fun decorateDeletedMessage(
-        viewHolder: MessageDeletedViewHolder,
-        data: MessageListItem.MessageItem
-    )
 }
